editor-api: added new test to test the RecordingFileService in isolation (#759)
* editor-api: added new test to test the RecordingFileService in isolation
core: Added new abstract test class, for better reusability
presenter-api: adapted test class to match changes made in core

* editor-api: fixed timeline marker positioning after splitting a recording
editor-fx: fixed words, grammar and word positioning

Fixes all issues mentioned in #758

+++ lect-editor-api/src/test/java/org/lecturestudio/editor/api/view/SplitRecordingMockView.java
@@ -6,18 +6,18 @@ import org.lecturestudio.core.view.ConsumerAction;
 
 public class SplitRecordingMockView implements SplitRecordingView {
 	public Action onCloseAction;
-	public ConsumerAction<Interval<Integer>> onSubmitAction;
-	public Interval<Integer> beginning;
-	public Interval<Integer> end;
+	public ConsumerAction<Interval<Long>> onSubmitAction;
+	public Interval<Long> beginning;
+	public Interval<Long> end;
 
 	@Override
-	public void setIntervals(Interval<Integer> beginning, Interval<Integer> end) {
+	public void setIntervals(Interval<Long> beginning, Interval<Long> end) {
 		this.beginning = beginning;
 		this.end = end;
 	}
 
 	@Override
-	public void setOnSubmit(ConsumerAction<Interval<Integer>> action) {
+	public void setOnSubmit(ConsumerAction<Interval<Long>> action) {
 		this.onSubmitAction = action;
 	}
